Refactor module storage
Currently some generic data structures and functions are located in
storage::physical module. Other storage backend like MySQL also depends
on these data structures or functions. So this commits moves them into
an upper-level - the storage module itself, for more clear code
readability and better module hierachy.
