+++ src/core.rs
@@ -32,7 +32,8 @@ use crate::{
         barrier_aes_gcm,
         barrier_view::BarrierView,
         physical,
-        physical::{Backend as PhysicalBackend, BackendEntry as PhysicalBackendEntry},
+        Backend as PhysicalBackend,
+        BackendEntry as PhysicalBackendEntry,
     },
 };
 
@@ -435,7 +436,7 @@ mod test {
     use serde_json::Value;
 
     use super::*;
-    use crate::storage::{barrier_aes_gcm, physical};
+    use crate::storage;
 
     #[test]
     fn test_core_init() {
@@ -448,8 +449,8 @@ mod test {
         let mut conf: HashMap<String, Value> = HashMap::new();
         conf.insert("path".to_string(), Value::String(dir.to_string_lossy().into_owned()));
 
-        let backend = physical::new_backend("file", &conf).unwrap();
-        let barrier = barrier_aes_gcm::AESGCMBarrier::new(Arc::clone(&backend));
+        let backend = storage::new_backend("file", &conf).unwrap();
+        let barrier = storage::barrier_aes_gcm::AESGCMBarrier::new(Arc::clone(&backend));
         let router = Arc::new(Router::new());
         let mounts = MountTable::new();
         let core = Arc::new(RwLock::new(Core {
@@ -499,8 +500,8 @@ mod test {
         let mut conf: HashMap<String, Value> = HashMap::new();
         conf.insert("path".to_string(), Value::String(dir.to_string_lossy().into_owned()));
 
-        let backend = physical::new_backend("file", &conf).unwrap();
-        let barrier = barrier_aes_gcm::AESGCMBarrier::new(Arc::clone(&backend));
+        let backend = storage::new_backend("file", &conf).unwrap();
+        let barrier = storage::barrier_aes_gcm::AESGCMBarrier::new(Arc::clone(&backend));
 
         let core = Arc::new(RwLock::new(Core { physical: backend, barrier: Arc::new(barrier), ..Default::default() }));
