+++ src/modules/pki/mod.rs
@@ -161,7 +161,7 @@ mod test {
     use crate::{
         core::{Core, SealConfig},
         logical::{Operation, Request},
-        storage::{barrier_aes_gcm, physical},
+        storage,
     };
 
     const CA_CERT_PEM: &str = r#"
@@ -1590,9 +1590,9 @@ xxxxxxxxxxxxxx
         let mut conf: HashMap<String, Value> = HashMap::new();
         conf.insert("path".to_string(), Value::String(dir.to_string_lossy().into_owned()));
 
-        let backend = physical::new_backend("file", &conf).unwrap();
+        let backend = storage::new_backend("file", &conf).unwrap();
 
-        let barrier = barrier_aes_gcm::AESGCMBarrier::new(Arc::clone(&backend));
+        let barrier = storage::barrier_aes_gcm::AESGCMBarrier::new(Arc::clone(&backend));
 
         let c = Arc::new(RwLock::new(Core { physical: backend, barrier: Arc::new(barrier), ..Default::default() }));
